openExp/requirements.txt
STOUT-pypi
pubchempy
pandas
rdkit
paragraph2actions
python-Levenshtein